removed unecasary(i think?) float cast

+++ pymc_marketing/clv/utils.py
@@ -350,4 +350,4 @@ def clv_summary(
         )
         summary_columns.append("monetary_value")
 
-    return customers[summary_columns].astype(float).reset_index()
+    return customers[summary_columns].reset_index()
